backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.2
supabase==2.7.4
google-generativeai==0.7.2
pdfplumber==0.11.4
pypdf==4.3.1
python-docx==1.1.2
pydantic==2.8.2
pydantic-settings==2.4.0
aiofiles==24.1.0
